LayoutTests/imported/w3c/web-platform-tests/fetch/private-network-access/non-secure-context.window-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'treat-as-public-address'.


PASS Local non secure page fetches local page.
FAIL Public non secure page fetches local page. assert_equals: expected "failure" but got "success"
